chore: #94 move block setters to separate file
